Update dependencies and supported browsers
I took a look at the analytics and found the majority of users are using
very modern browsers. So, I updated all of the browsers in babelrc for
the env preset (along with updating the preset itself for new data), and
noticed that all of the supported browsers support async functions
natively! Therefore, I don't need to include a polyfill for it anymore!
:tada: